Refactor column PHP type (#315)

+++ tests/Type/FloatTest.php
@@ -39,7 +39,7 @@ final class FloatTest extends TestCase
         $tableSchema = $db->getTableSchema('float_default');
 
         $this->assertSame('float', $tableSchema?->getColumn('Myfloat')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Myfloat')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Myfloat')->getPhpType());
         $this->assertSame(2.2300000000000001e-308, $tableSchema?->getColumn('Myfloat')->getDefaultValue());
 
         $db->createCommand()->insert('float_default', [])->execute();
@@ -86,7 +86,7 @@ final class FloatTest extends TestCase
         $tableSchema = $db->getTableSchema('float_default');
 
         $this->assertSame('float', $tableSchema?->getColumn('Myfloat')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Myfloat')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Myfloat')->getPhpType());
         $this->assertSame(2.2300000000000001e-308, $tableSchema?->getColumn('Myfloat')->getDefaultValue());
 
         $command = $db->createCommand();
