+++ tests/Type/NumericTest.php
@@ -39,7 +39,7 @@ final class NumericTest extends TestCase
         $tableSchema = $db->getTableSchema('numeric_default');
 
         $this->assertSame('numeric(38,0)', $tableSchema?->getColumn('Mynumeric')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Mynumeric')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Mynumeric')->getPhpType());
         $this->assertSame(38, $tableSchema?->getColumn('Mynumeric')->getSize());
         $this->assertSame(9.9999999999999998e+037, $tableSchema?->getColumn('Mynumeric')->getDefaultValue());
 
@@ -87,7 +87,7 @@ final class NumericTest extends TestCase
         $tableSchema = $db->getTableSchema('numeric_default');
 
         $this->assertSame('numeric(38,0)', $tableSchema?->getColumn('Mynumeric')->getDbType());
-        $this->assertSame('double', $tableSchema?->getColumn('Mynumeric')->getPhpType());
+        $this->assertSame('float', $tableSchema?->getColumn('Mynumeric')->getPhpType());
         $this->assertSame(38, $tableSchema?->getColumn('Mynumeric')->getSize());
         $this->assertSame(9.9999999999999998e+037, $tableSchema?->getColumn('Mynumeric')->getDefaultValue());
